django-app-main/backend_api/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
pymongo==3.12.3
pytz==2023.3.post1
sqlparse==0.2.4
typing-extensions==4.8.0
